tt-metal-llama3-70b/requirements.txt
# inference server requirements
pyjwt==2.7.0
flask==3.0.2
gunicorn==21.2.0
requests==2.31.0
# llama2 requirements
fairscale
fire
sentencepiece
blobfile
# tt-metal impl requirements
torch==2.2.1.0+cpu
transformers==4.38.0
tqdm==4.66.3
tiktoken==0.3.3
pytest==7.2.2